Thinhtran

+++ src/main/java/com/example/ecommerceweb/repository/BookRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BookRepository extends JpaRepository<Books, Integer> {
+public interface BookRepository extends JpaRepository<Books, Long> {
 
 }
